ButlerRobot/requirements.txt
wheel
robotframework>=6.0.2
robotframework-browser>=16.0.5
robotframework-seleniumlibrary
fastdeploy-python
pillow
imagehash
cryptography
datasets
pandas
openpyxl
pyautogui
rpaframework
